PullRequest: 6 <refactor> 重构了AbstracDataPipe
Merge branch refactor_datapipe_shanlu of git@gitlab.alipay-inc.com:oceanbase-cloud-platform/ob-data-mocker.git into 0.1.2_dev
https://code.alipay.com/oceanbase-cloud-platform/ob-data-mocker/pull_requests/6



* <refactor> DataPipe重构，区分了超时状态和管道关闭状态

* <fix> 修复单元测试不通过的问题

* <refactor> 增加sql异常时的回滚逻辑并且针对RandomGenerator提升性能

* <fix> 解决单元测试问题

* <fix> 解决单元测试不通过的问题

* <fix> 解决一个潜在的NPE问题

* <refactor> 修改Boolean到boolean，避免可能的npe问题

* <refactor> 优化MockBuffer

* <refactor> 回退对于MockBuffer的修改

* <refactor> 根据review意见修改代码

* <refactor> 增加字符编码值

+++ ob-mocker/src/main/java/com/oceanbase/tools/datamocker/core/write/SqlScriptWriter.java
@@ -105,7 +105,7 @@ public class SqlScriptWriter extends AbstractMockWriter {
     }
 
     @Override
-    protected Long doWrite(List<MockRowData> rows) throws IOException {
+    protected long doWrite(List<MockRowData> rows) throws IOException {
         MockRowData firstRow = rows.get(0);
         Set<String> columnSet = firstRow.columnNames();
         List<String> columnList = new ArrayList<>(columnSet);
@@ -162,7 +162,7 @@ public class SqlScriptWriter extends AbstractMockWriter {
         }
         String result = String.join("", sqlList);
         this.manager.write(result.getBytes(), 0, result.getBytes().length, true);
-        return (long) rows.size();
+        return rows.size();
     }
 
     @Override
